refactor(okutrade): move Chains enum to utils
